protocol/stack.tex: Allow record lists in the record store
Currently go-ipfs has (in routing/dht/providers.go) in-memory handling
for the provider listing.  That works well enough, but it seems like
we'd want to store this sort of thing in the generic record store to
avoid duplicate record-store-like code.  The problem with records like
this is that they're keyed off the multihash for the provided object,
but the records themselves will be created and signed by multiple
providing nodes.  That means we can't store a single signature as the
record-store entry (which provider would sign it?).  This commit adds
a record-list object that addresses this case.

The record-list object has Merkle-links to signatures where the link
names are the IDs for the publishing nodes (e.g. the providers for the
provider-list case, or wanters in wantlists, etc.).  Each linked
signature would have a signed payload with data containing the claim.
For example:

  I, <publisher-ID>, can provide <multihash-of-provided-object>.

or

  I, <publisher-ID>, would like to hear about changes to
  <multihash-of-wanted-object>.

or some more easily maching-parsable version of similar claims.

The Merkle chain of the record would be:

  record
    <publisher-ID>: <signature-ID>
  ↓
  signature
    Key: <provider-ID>
    Signee: <providing-claim-ID>
    …
  ↓
  <providing-claim-id>
    Data: <claim-payload>

Nodes that had reason to trust each other (e.g. to not forge providing
claims or to properly currate a provider-list) wouldn't have to fetch
and verify the signed data.  Nodes that had no reason to trust each
other (e.g. fetching a provider list from an untrusted node or
receiving a new providing-claim from an untrusted node) should aquire
and check the signature before using the providing-claim for lookup or
adding it to a provider-list.  Since most nodes won't trust each
other, I expect the signature, providing-claim, public-key, etc.,
packets would be passed around in one optimistic-transmit block [1],
and probably be optimistically hosted on the same nodes that host the
provider-list itself for that purpose.

[1]: https://github.com/ipfs/specs/pull/7#discussion_r31356057

License: MIT
Signed-off-by: W. Trevor King <wking@tremily.us>