modules: Drop boolean return from PublicKey.Verify
The only space between "this is invalid, but otherwise ok" and "we've
had an error, and can't figure out if the signature is ok or not"
seems to be known implementation shortcomings (e.g. "we don't handle
8192-bit RSA keys").  I think that's a pretty small fraction of cases,
so I don't mind folding it into the error result.  The existing Go
Verify APIs return only a boolean [1,2] or only an error [3,4], so I
don't think there's a pressing need for us to be even more flexible.

[1]: https://golang.org/pkg/crypto/ecdsa/#Verify
[2]: https://golang.org/pkg/crypto/dsa/#Verify
[3]: https://golang.org/pkg/crypto/rsa/#VerifyPKCS1v15
[4]: https://golang.org/pkg/crypto/rsa/#VerifyPSS