modules: Add AllKeys to Datastore
Copy the AllKeys() method from the blockstore to the datastore.  You
can't implement the blockstore's version efficiently using the
existing datastore interface, so I don't know how you'd add it on top
of a generic datastore that lacked such a method.  We don't want to
encourage drilling down through layers and using what should be
internal implementation details.

The previous paragraph explained why we need an AllKeys() method in
the datastore.  We also need to expose AllKeys() in the blockstore
interface, so we can build garbage collection and similar logic on top
of the blockstore, without having to drill down to the datastore layer
to write those tools (see, for example, the IRC discussion from [1]
through [2]).

[1]: https://botbot.me/freenode/ipfs/2015-06-23/?msg=42683298&page=4
[2]: https://botbot.me/freenode/ipfs/2015-06-23/?msg=42688156&page=4