modules: Reduce key/data indirection
Use Multihash and []byte directly, because they're designed to be as
generic as we'll need.  The initial motivation for this commit was
avoiding calling a passed-in Block.Keys in the Put implementation.  I
guess I don't really mind if we want a Blocks.Key alias for Multihash
and a Blocks.Block alias for []byte, but I don't think that gives us
much extra future-proofing, and I do see it giving us a bit of a
headache with casting, so it seemed better to just use the native
types.