docs: adds spec sections for mode and mtime metadata
I didn't include the pros and cons of the other options because I
don't think they should be included in the spec.  That is, if I'm
implementing this, I read the spec to understand what I should
implement and how it should behave, I don't want to read a bunch
of exposition on how we got to where we are.

The decision log is important though, should it go in
`ipfs/notes`?  Or maybe we could have accompanying `*-notes.md`
files?

Follows on from/supersedes #223