Describe the double-hashed FindProviders reframe request in parallel to the current method.
FindHashedSHA256Request requests content routing records using a derived hash of a CID of interest.

This extension to reframe also introduces authenticaled content routing records that can be
returned by this method. These records will blind the true peer ID of the provider, and
are extended with a signature allowing the requester to validate their authenticity.