feat: IPIP 298 (allow|deny)list for IPFS Nodes and Gateways
Spec for the IPIP-298.

It adds a standard reference for content list definition, should these
be allow lists or deny lists. It supports CIDv1, sha256 hashed CIDv1,
and absolute and partial path matching.
It is amied at offering a common schema for node operators to interact
with content lists.

It should be noted this IPIP was first written by @foreseaz, and this PR
is an iteration over their work in #299.